+++ Source/Frette/Public/Components/BodyPart/FretteBodyPartInstance.h
@@ -50,7 +50,7 @@ private:
 	void RemoveGameplayAbilities(const TArray<TSubclassOf<UGameplayAbility>>& Abilities) const;
 	void BuildRuleLookup();
 	void SetMinDeltaValueThreshold();
-	float ClampDelta(int Value, FGameplayTag Tag);
+	int ClampDelta(int Value, FGameplayTag Tag);
 	void CheckDeltaRules(FGameplayTag Tag, FFretteBodyPartContext& Context, int ClampedDelta) const;
 	void CheckAccumulatedValueRules(FGameplayTag Tag, FFretteBodyPartContext& Context, int CurrentValue) const;
 	const TArray<FFretteEffectRuleEntry>* GetRulesForEvent(const EBodyPartEventType EventType, const FGameplayTag& EffectTag) const;
